[rend2] Fixing some variable size related compiler warnings

+++ shared/rd-rend2/tr_backend.cpp
@@ -462,7 +462,7 @@ void GL_DrawIndexed(
 		GLenum primitiveType,
 		int numIndices,
 		GLenum indexType,
-		int offset,
+		size_t offset,
 		int numInstances,
 		int baseVertex)
 {
